Sort order, shorten base36 and login key
The sort order was broken, sort -r doesn't work. Now sorting of
links is done in reverse order in the log, or when they are entered
or modified.

Shorten was broken when trying to calculate the next base 36 number.
When going from z to 10, z kept coming up in the sort order as the next
rather than 10 which would increment to 11. Added padding to the number
fixes it.

Login key management was fixed to resolve issue of not allowing the same
user from multiple systems. Now key is removed only in expiration and
logout.

Resolves #0027, #0028, #0023